Adds more functionalities

+++ django/stu/views.py
@@ -10,8 +10,10 @@ from django.views.generic.edit import CreateView, UpdateView, DeleteView
 
 def student_index(request):
 	# student = Student.objects.get(pk=3)
-	if request.session.get('student_id',None)==None:
+	if request.session.get('student_id','None')=='None':
+		# print(request.session.get('student_id','None'))
 		return redirect('stu:login')
+	# print(request.session.get('student_id','None'))
 
 	student = Student.objects.get(pk=request.session['student_id'])
 	applications = student.application_set.all()
@@ -39,6 +41,8 @@ def company_details(request, com_name):
 # return HttpResponse("Hello, world. You're at the polls index.")
 
 def apply(request, job_id, stu_id):
+	if request.session.get('student_id',None)==None:
+		return redirect('stu:login')
 	app = Application.objects.create(student_id=Student.objects.get(pk=stu_id), job_id=Job.objects.get(pk=job_id),
 									 status='Applied')
 	# print(app)
@@ -49,35 +53,67 @@ def apply(request, job_id, stu_id):
 # return redirect('stu_index')
 
 def check(request):
-	user = Student.objects.get(email=request.POST['email'])
-	if user.password == request.POST['password']:
+	try:
+		user = Student.objects.get(email=request.POST['email'])
+	except Student.DoesNotExist:
+		user = None
+
+	if user!= None and user.password == request.POST['password']:
 		request.session['student_id'] = user.id
 		return redirect('stu:home')
-	return HttpResponse("<p>login error</p>")
+
+	try:
+		user = Company.objects.get(email=request.POST['email'])
+	except Company.DoesNotExist:
+		user = None
+
+	if user!=None and user.password == request.POST['password']:
+		request.session['company_id'] = user.id
+		return redirect('stu:company_home')
+	request.session['login_error']='incorrect credentials'
+	return redirect('stu:login')
 	# return redirect('stu:login')
 
 def login(request):
-	return render(request, 'company/login_form.html')
+	msg=''
+	if request.session.get('login_error',None)!=None:
+		msg=request.session.get('login_error')
+		del request.session['login_error']
+
+	return render(request, 'company/login_form.html',{'msg':msg})
 
 def logout(request):
-	del request.session['student_id']
-	request.session.modified = True
-	return redirect('stu:login', permanent=True)
+	# print('None')
+	if type(request.session.get('student_id',None))==int:
+		del request.session['student_id']
+		request.session.modified = True
+	if not request.session.get('company_id',None)==None:
+		del request.session['company_id']
+		request.session.modified = True
+	return redirect('stu:login')
 
 
 def company_index(request):
-	company = Company.objects.get(pk=2)
+	if request.session.get('company_id',None)==None:
+		return redirect('stu:login')
+	# company = Company.objects.get(pk=2)
+	company = Company.objects.get(pk=request.session['company_id'])
 	jobs = company.job_set.all()
 	return render(request, 'company/index.html', {'jobs': jobs})
 
 def job_delete(request, job_id):
+	if request.session.get('company_id',None)==None:
+		return redirect('stu:login')
+
 	Job.objects.filter(id=job_id).delete()
 	return redirect('stu:company_home')
 
 def job_add(request):
+	if request.session.get('company_id',None)==None:
+		return redirect('stu:login')
 
 	# replace Company.objectsget(id=2) by company's session variable
-	add = Job.objects.create(pos=request.POST.get('pos',''),company=Company.objects.get(id=2),skills=request.POST.get('skills',''))
+	add = Job.objects.create(pos=request.POST.get('pos',''),company=Company.objects.get(id=request.session.get('company_id')),skills=request.POST.get('skills',''))
 	return redirect('stu:company_home')
 
 class JobCreate(CreateView):
@@ -85,12 +121,12 @@ class JobCreate(CreateView):
 	model = Job
 	fields = ['pos', 'skills']
 
-class CompanyView(generic.ListView):
-	template_name = 'company/index.html'
-	context_object_name = 'jobs'
-	def get_queryset(self):
-		company = Company.objects.get(pk=2)
-		return company.job_set.all()
+# class CompanyView(generic.ListView):
+# 	template_name = 'company/index.html'
+# 	context_object_name = 'jobs'
+# 	def get_queryset(self):
+# 		company = Company.objects.get(pk=2)
+# 		return company.job_set.all()
 
 # class LogIn(CreateView):
 #     template_name = "company/login_form.html"
